refactor(audit): resolve structural findings (#5558)
